+++ vice/src/c64/cart/silverrock128.c
@@ -218,7 +218,7 @@ static void silverrock128_io1_store(WORD addr, BYTE value)
     } else {
         if ((addr >= 0x0) && (addr <= 0x0f)) {
             /* Cartridge HW rev02 */
-            bank_number = addr;
+            bank_number = (BYTE)addr;
             /* safe check that we've done things right... */
             bank_index = ((value & 0xf0) >> 4);
             if (bank_number != bank_seq[bank_index]) {
