+++ vice/src/mouse.c
@@ -410,10 +410,10 @@ BYTE mouse_poll(void)
         /* calculate the interval between the latest two mousedrv
          * updates in emulated cycles */
         os_iv = os_now - latest_os_ts;
-        if (os_iv > vsyncarch_frequency()) {
-            os_iv = vsyncarch_frequency(); /* more than a second response time?! */
+        if (os_iv > (unsigned long)vsyncarch_frequency()) {
+            os_iv = (unsigned long)vsyncarch_frequency(); /* more than a second response time?! */
         }
-        emu_iv = os_iv * emu_units_per_os_units;
+        emu_iv = os_iv * (unsigned long)emu_units_per_os_units;
 #ifdef DEBUG_MOUSE
         log_message(mouse_log,
                     "New interval os_now %lu, os_iv %lu, emu_iv %lu",
